+++ src/analyze/+mag/+process/Filter.m
@@ -8,10 +8,10 @@ classdef Filter < mag.process.Step
     end
 
     properties
-        % ONRANGECHANGE How long to remove when range changes.
-        OnRangeChange (1, 2) duration
+        % ONRANGECHANGE How many vectors to remove when range changes.
+        OnRangeChange (1, 2) {mustBeA(OnRangeChange, ["double", "duration"])} = zeros(1, 2)
         % ONMODECHANGE How many vectors to remove when mode changes.
-        OnModeChange (1, 2) double = zeros(1, 2)
+        OnModeChange (1, 2) {mustBeA(OnModeChange, ["double", "duration"])} = zeros(1, 2)
     end
 
     methods
@@ -37,15 +37,15 @@ classdef Filter < mag.process.Step
 
             value = this.Description + " After said events, onboard filtering " + ...
                 "needs time to adjust, thus some data points are dropped for display purposes. " + ...
-                "For range changes, " + string(this.OnRangeChange(1)) + "-worth before and " + string(this.OnRangeChange(2)) + ...
-                "-worth after are dropped, and for mode changes, " + string(this.OnModeChange(1)) + " vector before and " + ...
+                "For range changes, " + string(this.OnRangeChange(1)) + " before and " + string(this.OnRangeChange(2)) + ...
+                " after are dropped, and for mode changes, " + string(this.OnModeChange(1)) + " before and " + ...
                 string(this.OnModeChange(2)) + " after are dropped.";
         end
 
         function data = apply(this, data, ~)
 
             arguments
-                this
+                this (1, 1) mag.process.Filter
                 data timetable
                 ~
             end
@@ -57,21 +57,12 @@ classdef Filter < mag.process.Step
 
             % Filter data points at mode changes.
             if ~isequal(this.OnModeChange, zeros(1, 2))
-
-                locMode = [true; diff(events.DataFrequency) ~= 0];
-
-                for t = events.Time(locMode)'
-
-                    idxTime = find(events.Time == t);
-                    data(idxTime + (this.OnModeChange(1):this.OnModeChange(2)), :) = [];
-                end
+                data = this.cropDataWithRange(events, data, "DataFrequency", this.OnModeChange);
             end
 
             % Filter duration at range changes.
-            locRange = [true; diff(events.Range) ~= 0];
-
-            for t = events.Time(locRange)'
-                data(timerange(t + this.OnRangeChange(1), t + this.OnRangeChange(2), "closed"), :) = [];
+            if ~isequal(this.OnRangeChange, zeros(1, 2))
+                data = this.cropDataWithRange(events, data, "Range", this.OnRangeChange);
             end
 
             % Filter out between config and ramp mode.
@@ -83,40 +74,24 @@ classdef Filter < mag.process.Step
             if (nnz(locConfig) == 2) && any(contains([events.Label(idxConfig(1):idxConfig(end))], "Ramp"))
                 data(timerange(events.Time(idxConfig(1)), events.Time(idxConfig(end)), "closed"), :) = [];
             end
-
-            % Make sure no sliced packets remain.
-            % data = this.removeSlicedSequences(data);
         end
     end
 
     methods (Static, Access = private)
 
-        function data = removeSlicedSequences(data)
+        function data = cropDataWithRange(events, data, name, range)
 
-            events = data.Properties.Events;
-            modeEvents = events(~ismissing(events.Duration), :);
+            locEvent = [true; diff(events.(name)) ~= 0];
 
-            locCombine = ([NaN; diff(modeEvents.DataFrequency)] == 0);
-            modeEvents(locCombine, :) = [];
+            for t = events.Time(locEvent)'
 
-            for e = 1:height(modeEvents)
-
-                if e == height(modeEvents)
-                    endTime = data.t(end);
+                if isa(range, "duration")
+                    data(timerange(t + range(1), t + range(2), "closed"), :) = [];
                 else
-                    endTime = modeEvents.Time(e + 1);
+
+                    idxTime = find(events.Time == t);
+                    data(idxTime + (range(1):range(2)), :) = [];
                 end
-
-                locTime = isbetween(data.t, modeEvents.Time(e), endTime, "closedleft");
-                vectorsPerPacket = modeEvents.DataFrequency(e) * modeEvents.PacketFrequency(e);
-
-                sequence = mag.process.Step.correctSequence(data.sequence);
-                [~, idxSequence] = unique(sequence);
-
-                idxPeriod = idxSequence(locTime(idxSequence));
-                locSequence = diff(idxPeriod) ~= vectorsPerPacket;
-
-                data(ismember(sequence, sequence(idxPeriod(locSequence))), :) = [];
             end
         end
     end